+++ src/App.tsx
@@ -9,11 +9,16 @@ import { HomePage, Login, Register } from "./pages"
 import { Navbar } from "./components/Navbar/Navbar"
 import { AdminModal, AuthModal } from "./components/ui/Modals"
 import { useModals } from "./store/ui/useModals"
+import useUserCartStore from "./store/user/userCartStore"
+import supabase from "./utils/supabaseClient"
+import useUserStore from "./store/user/userStore"
 
 function App() {
   const { isOpen, closeModal } = useModals()
 
   const darkMode = useDarkMode()
+  const userCartStore = useUserCartStore()
+  const userStore = useUserStore()
 
   useEffect(() => {
     if (darkMode.isDarkMode) {
@@ -24,6 +29,15 @@ function App() {
       document.getElementsByTagName("html")[0].classList.add("light")
     }
   }, [darkMode])
+
+  useEffect(() => {
+    if (userStore.userId) {
+      const fetchUserCart = async () => {
+        const response = await supabase.from("users_cart").select("products").eq("id", userStore.userId)
+      }
+      fetchUserCart()
+    }
+  }, [userCartStore.products])
 
   return (
     <div
