+++ src/store/user/userCartStore.ts
@@ -4,37 +4,25 @@ import supabase from "../../utils/supabaseClient"
 
 export interface IProduct {
   id: string
-  label: string
+  title: string
+  sub_title: string
   price: number
   img_url: string
   quantity: number
+  on_stock: number
 }
 
 interface UserCartStore {
   products: IProduct[]
   cartQuantity: number
   setCartQuantityFromDB: (cartQuantity: number) => void
-  increaseItemQuantity: (id: string) => void
+  increaseProductQuantity: (product: IProduct) => void
   decreaseItemQuantity: (id: string) => void
   setItemQuantity0: (id: string) => void
 }
 
-export const increaseItemQuantityinDB = async (id: string) => {
-  //db manipulations here
-  const { data } = await supabase.from("products").select("id, label, price, img_url").eq("id", id)
-
-  return data && data.length > 0
-    ? {
-        id: data[0].id,
-        label: data[0].label,
-        price: data[0].price,
-        img_url: data[0].img_url,
-        quantity: 1, //logic
-      }
-    : null
-}
-
 export const decreaseItemQuantity = (products: IProduct[], id: string): IProduct[] => {
+  console.log(39, "decreaseItemQuantity - ")
   return products.map(product => {
     if (product.id === id) {
       return {
@@ -47,12 +35,15 @@ export const decreaseItemQuantity = (products: IProduct[], id: string): IProduct
 }
 
 export const setCartQuantityFromDB = (cartQuantity: number) => {
+  console.log(52, "setCartQuantityFromDB")
   //I don't do backend request here because I want useEffect to fire this function
   //I mean I don't call this function on click
   return cartQuantity
 }
 
 export const setItemQuantity0 = (products: IProduct[], id: string): IProduct[] => {
+  console.log(59, "setItemQuantity0")
+
   return products.map(product => {
     if (product.id === id) {
       return {
@@ -64,19 +55,23 @@ export const setItemQuantity0 = (products: IProduct[], id: string): IProduct[] =
   })
 }
 
+// export const setUserCartFromDB = (products: IProduct[], cartQuantity: number): IProduct[] => {
+//   console.log(75, "setUserCartFromDB")
+
+//   return {}
+// }
+
 type SetState = (fn: (prevState: UserCartStore) => UserCartStore) => void
 
 const userCartStore = (set: SetState): UserCartStore => ({
   products: [],
   cartQuantity: 0,
-  async increaseItemQuantity(id: string) {
-    const newProduct = await increaseItemQuantityinDB(id)
-    if (newProduct)
-      set((state: UserCartStore) => ({
-        ...state,
-        products: [...state.products, newProduct],
-        cartQuantity: this.cartQuantity + 1,
-      }))
+  increaseProductQuantity(product: IProduct) {
+    set((state: UserCartStore) => ({
+      ...state,
+      products: [...state.products, product],
+      cartQuantity: state.cartQuantity + 1,
+    }))
   },
   decreaseItemQuantity(id: string) {
     set((state: UserCartStore) => ({
